pytorch_segmentation/fcn/requirements.txt
python>=3.8.0
numpy==1.22.0
torch==1.12.1
pillow==8.4.0
torchvision==0.13.1
